+++ dprint.json
@@ -20,13 +20,13 @@
     "apps/oxlint/src-js/bindings.js",
     "apps/oxlint/src-js/bindings.d.ts",
     "napi/{transform,minify,playground}/index.js",
-    "napi/{parser,transform,minify,playground}/index.d.ts",
-    "napi/{parser,transform,minify,playground}/*.wasi-browser.js",
-    "napi/{parser,transform,minify,playground}/*.wasi.cjs",
-    "napi/{parser,transform,minify,playground}/wasi-worker-browser.mjs",
-    "napi/{parser,transform,minify,playground}/wasi-worker.mjs",
-    "napi/{parser,transform,minify,playground}/browser.js",
-    "napi/parser/bindings.mjs",
+    "napi/{parser,transform,minify,playground}/**/index.d.ts",
+    "napi/{parser,transform,minify,playground}/**/*.wasi-browser.js",
+    "napi/{parser,transform,minify,playground}/**/*.wasi.cjs",
+    "napi/{parser,transform,minify,playground}/**/wasi-worker-browser.mjs",
+    "napi/{parser,transform,minify,playground}/**/wasi-worker.mjs",
+    "napi/{parser,transform,minify,playground}/**/browser.js",
+    "napi/parser/src-js/bindings.mjs",
     "npm/*/package.json",
     "npm/oxlint/configuration_schema.json",
     "npm/oxc-wasm/**",
