+++ napi/parser/package.json
@@ -2,15 +2,15 @@
   "name": "oxc-parser",
   "version": "0.90.0",
   "type": "commonjs",
-  "main": "index.mjs",
-  "browser": "wasm.mjs",
+  "main": "src-js/index.mjs",
+  "browser": "src-js/wasm.mjs",
   "scripts": {
-    "build-dev": "napi build --esm --platform --js bindings.mjs",
+    "build-dev": "napi build --esm --platform --js bindings.mjs --output-dir src-js",
     "build-test": "pnpm run build-dev --profile coverage",
     "build": "pnpm run build-dev --features allocator --release",
     "postbuild-dev": "node patch.mjs",
     "build-wasi": "pnpm run build-dev --release --target wasm32-wasip1-threads",
-    "build-npm-dir": "rm -rf npm-dir && napi create-npm-dirs --npm-dir npm-dir && pnpm napi artifacts --npm-dir npm-dir --output-dir .",
+    "build-npm-dir": "rm -rf npm-dir && napi create-npm-dirs --npm-dir npm-dir && pnpm napi artifacts --npm-dir npm-dir --output-dir src-js",
     "build-browser-bundle": "node build-browser-bundle.mjs",
     "test": "tsc && pnpm run test-node run",
     "test-node": "vitest --dir ./test",
@@ -38,25 +38,8 @@
     "url": "https://github.com/sponsors/Boshen"
   },
   "files": [
-    "index.d.ts",
-    "index.mjs",
-    "wrap.mjs",
-    "wasm.mjs",
-    "bindings.mjs",
-    "webcontainer-fallback.js",
-    "generated/constants.mjs",
-    "generated/deserialize/js.mjs",
-    "generated/deserialize/ts.mjs",
-    "generated/lazy/constructors.mjs",
-    "generated/lazy/types.mjs",
-    "generated/lazy/walk.mjs",
-    "raw-transfer/common.mjs",
-    "raw-transfer/eager.mjs",
-    "raw-transfer/lazy.mjs",
-    "raw-transfer/lazy-common.mjs",
-    "raw-transfer/node-array.mjs",
-    "raw-transfer/supported.mjs",
-    "raw-transfer/visitor.mjs"
+    "generated",
+    "src-js"
   ],
   "publishConfig": {
     "registry": "https://registry.npmjs.org/",
@@ -100,6 +83,6 @@
         "fs": false
       }
     },
-    "dtsHeaderFile": "header.js"
+    "dtsHeaderFile": "src-js/header.js"
   }
 }
