+++ napi/parser/src-js/index.d.ts
