+++ napi/parser/src-js/parser.wasi-browser.js
