+++ napi/parser/src-js/raw-transfer/common.mjs
@@ -1,10 +1,10 @@
 import os from 'node:os';
+import { BUFFER_ALIGN, BUFFER_SIZE, IS_TS_FLAG_POS } from '../../generated/constants.mjs';
 import {
   getBufferOffset,
   parseAsyncRaw as parseAsyncRawBinding,
   parseSyncRaw as parseSyncRawBinding,
 } from '../bindings.mjs';
-import { BUFFER_ALIGN, BUFFER_SIZE, IS_TS_FLAG_POS } from '../generated/constants.mjs';
 import { rawTransferSupported } from './supported.mjs';
 
 // Throw an error if running on a platform which raw transfer doesn't support.
