+++ napi/parser/src-js/raw-transfer/lazy-common.mjs
