+++ napi/parser/src-js/wasi-worker-browser.mjs
