+++ napi/parser/src-js/wasi-worker.mjs
