+++ napi/parser/src-js/webcontainer-fallback.js
