+++ napi/parser/src-js/wrap.mjs
