+++ napi/parser/test-browser/parse.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from 'vitest';
-import { parseAsync, parseSync } from '../wasm.mjs';
+import { parseAsync, parseSync } from '../src-js/wasm.mjs';
 
 test('parseSync', () => {
   const result = parseSync('test.js', 'ok');
