+++ napi/parser/test/lazy-deserialization.test.ts
@@ -4,7 +4,7 @@
 
 import { describe, expect, it } from 'vitest';
 
-import { parseSync } from '../index.mjs';
+import { parseSync } from '../src-js/index.mjs';
 
 function parseSyncLazy(filename, code, options = null) {
   return parseSync(filename, code, { ...options, experimentalLazy: true });
