+++ napi/parser/test/worker.mjs
@@ -1,2 +1,2 @@
-import { parseSync } from '../index.mjs';
+import { parseSync } from '../src-js/index.mjs';
 parseSync('test.js', '');
